refactor(#44) databaseを具体的なpostgresへと修正
